% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_alluvial.R
\name{data_sort}
\alias{data_sort}
\title{Sorts a dataframe.}
\usage{
data_sort(
  data,
  cols,
  wt = NULL,
  method = c("tsp", "neighbornet", "greedy_wolf", "greedy_wblf", "none", "random"),
  column_method = c("tsp", "neighbornet", "none", "random"),
  weight_scalar = 5e+05,
  fixed_column = NULL,
  verbose = FALSE,
  options = NULL
)
}
\arguments{
\item{data}{A data frame/tibble. Must be in one of two formats:
(1) wt == NULL: Each row represents an entity, each column represents a grouping, and each entry represents the membership of the entity in that row to the grouping in that column. Must contain at least two columns (two cols).
(2) wt != NULL: Each row represents a combination of groupings, each column from \code{cols} represents a grouping, and the column \code{wt} represents the number of entities in that combination of groupings. Must contain at least three columns (two \code{cols}, one \code{wt}).}

\item{cols}{Optional character vector. Vector of column names from \code{data} to be used in graphing (i.e., alluvial plotting). Mutually exclusive with \code{column1} and \code{column2}.}

\item{wt}{Optional character. Column name from \code{data} that contains the weights of each combination of groupings if \code{data} is in format (2) (see above).}

\item{method}{Character. Algorithm with which to sort the values in the dataframe. Can choose from: 'tsp', 'greedy_wolf', 'greedy_wblf', 'none'. 'tsp' performs Traveling Salesman Problem solver from the TSP package. greedy_wolf' implements a custom greedy algorithm where one layer is fixed, and the other layer is sorted such that each node is positioned as close to its largest parent from the fixed side as possible in a greedy fashion. 'greedy_wblf' implements the 'greedy_wolf' algorithm described previously twice, treating each column as fixed in one iteration and free in the other iteration. 'greedy_wolf' and 'greedy_wblf' are only valid when \code{cols} has exactly two entries. 'random' randomly maps blocks. 'none' keeps the mappings as-is when passed into the function.}

\item{column_method}{Character. Algorithm to use for determining column order. Options are 'tsp' (default), 'random', and 'none'.}

\item{weight_scalar}{Positive integer. Scalar with which to multiply edge weights after taking their -log in the distance matrix for nodes with a nonzero edge. Only applies when \code{method == 'tsp'}.}

\item{fixed_column}{Character or Integer. Name or position of the column in \code{cols} to keep fixed during sorting. Only applies when \code{method == 'greedy_wolf'}.}

\item{verbose}{Logical. If TRUE, will display messages during the function.}

\item{options}{Additional arguments. See data_sort_options}

\item{output_df_path}{Optional character. Output path for the output data frame, in rds format. If \code{NULL}, then will not be saved.}
}
\value{
A data frame where each row represents an alluvium and each column represents an axis. Each column of \code{cols} represents an axis, stored as a factor ordered in ascending order of strata. There is an additional column \code{wt} ('value' if NULL) that represents the size of the alluvium in that row. The order of columns represents the recommended order of axes.
}
\description{
Sorts a dataframe with the algorithm specified by \code{method}.
}
\examples{
# Example 1: data format 1
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
clus_df_gather <- data_sort(
    data,
    cols = c("method1", "method2"),
    method = "tsp",
    column_method = "tsp"
)

# Example 2: data format 2
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
clus_df_gather <- data |>
    dplyr::mutate_if(is.numeric, function(x) factor(x, levels = as.character(sort(unique(x))))) |>
    dplyr::group_by_all() |>
    dplyr::count(name = "value")
clus_df_gather <- data_sort(
    clus_df_gather,
    cols = c("method1", "method2"),
    wt = "value",
    method = "tsp",
    column_method = "tsp"
)

}
