% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_calculation.R
\name{determine_crossing_edges}
\alias{determine_crossing_edges}
\title{Determine overlapping edges}
\usage{
determine_crossing_edges(
  data,
  cols = NULL,
  wt = "value",
  weighted_metric = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame, tibble, or CSV file path. Must be in one of two formats:
(1) wt == NULL: Each row represents an entity, each column represents a grouping, and each entry represents the membership of the entity in that row to the grouping in that column. Must contain at least two columns (two cols).
(2) wt != NULL: Each row represents a combination of groupings, each column from \code{cols} represents a grouping, and the column \code{wt} represents the number of entities in that combination of groupings. Must contain at least three columns (two \code{cols}, one \code{wt}).}

\item{cols}{Optional character vector. Vector of column names from \code{data} to be used in graphing (i.e., alluvial plotting). Mutually exclusive with \code{column1} and \code{column2}.}

\item{wt}{Optional character. Column name from \code{data} that contains the weights of each combination of groupings if \code{data} is in format (2) (see above). If null, then sets \code{weighted_metric} to FALSE.}

\item{weighted_metric}{Logical. Determines if the objective is total number of edge crossings (weighted_metric=FALSE) or sum of product of overlapping edge weights (weighted_metric=TRUE).}

\item{verbose}{Logical. If TRUE, will display messages during the function.}
}
\value{
If return_weighted_layer_free_objective is FALSE (default): A list of values, as follows:
'lode_df': A data frame containing the following columns:
\itemize{
\item alluvium: A specific alluvium/edge.
\item count: The weight of the alluvium/edge.
\item x1, x2, ...: Each xi represents the x position of axis/layer i.
\item y1, y2, ...: Each yi represents the height of a lode in axis/layer i.
\item stratum1, stratum2, ...: Each stratumi represents the stratum through which the alluvial crosses in axis/layer i.
\item weight1: The weight of the first alluvium, corresponding to the 'count' column in \code{lode_df}.
\item weight2: The weight of the second alluvium, corresponding to the 'count' column in \code{lode_df}.
'output_objective': An integer representing the sum of products of overlapping edge weights.
}
}
\description{
Determine overlapping edges of k-partite graph.
}
\examples{
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
data <- sort_to_uncross(data, cols = c("method1", "method2"), method = "tsp")
result <- determine_crossing_edges(data, cols = c("method1", "method2"))

}
