% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R
\name{make_stratum_color_list}
\alias{make_stratum_color_list}
\title{Make stratum color list}
\usage{
make_stratum_color_list(data, cols, mapping)
}
\arguments{
\item{data}{A data frame with columns corresponding to axes, with factors indicating sorting. (eg run through data_sort)}

\item{cols}{Character vector. Vector of column names from \code{data} to be used in graphing (i.e., alluvial plotting).}

\item{mapping}{List. Output from data_color.}
}
\value{
A vector of colors.
}
\description{
Convert color mapping made by data_color into list for scale_fill_manual
}
\examples{
# Example 1
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
cols = c("method1", "method2")
clus_df_gather <- data_sort(data, cols = cols, method = "tsp")
color_mapping <- data_color(data = clus_df_gather, cols = cols)
color_list <- make_stratum_color_list(data = clus_df_gather, cols = cols, mapping = color_mapping)

}
