% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{plot_alluvial_internal}
\alias{plot_alluvial_internal}
\title{Just the plotting}
\usage{
plot_alluvial_internal(
  df,
  graphing_columns,
  column_weights,
  color_list = NULL,
  color_boxes = TRUE,
  color_bands = FALSE,
  color_band_list = NULL,
  color_band_column = NULL,
  color_band_boundary = FALSE,
  alluvial_alpha = 0.5,
  coloring_algorithm = "advanced",
  cutoff = 0.5,
  output_plot_path = NULL,
  color_val = NULL,
  include_labels_in_boxes = TRUE,
  include_axis_titles = FALSE,
  include_group_sizes = FALSE,
  verbose = FALSE,
  print_params = FALSE,
  box_width = 1/3,
  text_width = 1/4,
  min_text = 4,
  text_size = 14,
  auto_adjust_text = TRUE,
  axis_text_size = 2,
  axis_text_vjust = 0,
  save_height = 6,
  save_width = 6,
  dpi = 300,
  rasterise_alluvia = FALSE,
  keep_y_labels = FALSE,
  keep_x_labels = TRUE,
  box_line_width = 1,
  do_compute_alluvial_statistics = TRUE,
  coloring_algorithm_advanced_option = "leiden",
  resolution = 1,
  add_legend = FALSE,
  legend_loc = "right",
  flip_xy = FALSE
)
}
\arguments{
\item{df}{A data frame, tibble, or CSV file path. Must be in the format as the output of wompwomp::data_sort.}

\item{graphing_columns}{Character vector. Vector of column names from \code{df} to be used in graphing (i.e., alluvial plotting). Mutually exclusive with \code{column1} and \code{column2}.}

\item{column_weights}{Character. Column name from \code{df} that contains the weights of each combination of groupings if \code{df} is in format (2) (see above).}

\item{color_list}{Optional named list or vector of colors to override default group colors.}

\item{color_boxes}{Logical. Whether to color the strata/boxes (representing groups).}

\item{color_bands}{Logical. Whether to color the alluvia/edges (connecting the strata).}

\item{color_band_list}{Optional named list or vector of colors to override default band colors.}

\item{color_band_column}{Optional Character. Which column to use for coloring bands.}

\item{color_band_boundary}{Logical. Whether or not to color boundaries between bands}

\item{alluvial_alpha}{Numeric between 0 and 1. Transparency level for the alluvial bands.}

\item{coloring_algorithm}{Character. Matching colors methods. Choices are 'advanced' (default), 'none', 'left', 'right', or any value in \code{graphing_columns}.}

\item{cutoff}{Numeric If \code{coloring_algorithm != 'none' and coloring_algorithm != 'advanced'}, sets the cutoff for color matching, below which a new color will be assigned.}

\item{output_plot_path}{Character. File path to save the plot (e.g., "plot.png"). If \code{NULL}, then will not be saved.}

\item{color_val}{Optional named list where the entries are colors and the names correspond to values of the dataframe that should use those colors}

\item{include_labels_in_boxes}{Logical. Whether to include text labels inside the rectangular group boxes.}

\item{include_axis_titles}{Logical. Whether to display axis titles for column1 and column2.}

\item{include_group_sizes}{Logical. If \code{TRUE}, includes group sizes in the labels (e.g., "Group A (42)").}

\item{verbose}{Logical. If TRUE, will display messages during the function.}

\item{print_params}{Logical. If TRUE, will print function params.}

\item{box_width}{Numeric between 0 and 1. Box width}

\item{text_width}{Numeric between 0 and 1. Text width}

\item{min_text}{Integer greater than 0. Min text}

\item{text_size}{Integer greater than 0. Text size (works whether auto_adjust_text is TRUE or FALSE).}

\item{auto_adjust_text}{Logical. Whether to automatically adjust text size to fit in box.}

\item{axis_text_size}{Integer greater than 0. Axis text size}

\item{axis_text_vjust}{Integer. Axis text vjust}

\item{save_height}{Integer greater than 0. Save height, in inches}

\item{save_width}{Integer greater than 0. Save width, in inches}

\item{dpi}{Integer greater than 0. DPI for \code{output_plot_path}, if \code{output_plot_path} is a raster image or \code{rasterise_alluvia} is TRUE}

\item{rasterise_alluvia}{Logical. Whether to rasterize the alluvia if \code{output_plot_path} is a PDF. Can save space if DPI low enough}

\item{keep_y_labels}{Keep y labels}

\item{keep_x_labels}{Keep x labels}

\item{box_line_width}{Box line width}

\item{do_compute_alluvial_statistics}{Internal flag; not recommended to modify.}

\item{coloring_algorithm_advanced_option}{Character. If \code{coloring_algorithm == 'advanced'}, then choose graph clustering algorithm. Choices are 'leiden' (default) or 'louvain'.}

\item{resolution}{Numeric If \code{coloring_algorithm == 'advanced'}, then choose resolution for the graph clustering algorithm. Affects coloring of both bands and boxes.}

\item{add_legend}{Logical. If TRUE, will generate a legend of the colors of boxes and alluvial}

\item{legend_loc}{Character. Location of legend. Only applies if \code{add_legened == TRUE}. Choices are 'right' (default), 'left', 'bottom', 'top'}

\item{flip_xy}{Logical. Flip x and y (rotate plot 90 degrees).}
}
\value{
A \code{ggplot2} object representing the alluvial plot.
}
\description{
Just the plotting.......
}
\examples{
# Example:
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
clus_df_gather <- df |>
    dplyr::mutate_if(is.numeric, function(x) factor(x, levels = as.character(sort(unique(x))))) |>
    dplyr::group_by_all() |>
    dplyr::count(name = "value")
clus_df_gather <- data_sort(
    clus_df_gather,
    graphing_columns = c("method1", "method2"),
    column_weights = "value",
    sorting_algorithm = "none"
)
p <- plot_alluvial_internal(
    clus_df_gather,
    graphing_columns = c("method1", "method2"),
    column_weights = "value",
    coloring_algorithm = "right"
)

}
