% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cli.R
\name{run_cli}
\alias{run_cli}
\title{Run CLI entry point for data preprocessing}
\usage{
run_cli(args)
}
\arguments{
\item{args}{A character vector of command-line arguments.}
}
\value{
Called for side effects. Invisibly returns \code{NULL}.
}
\description{
This function is called internally to handle command line arguments.
}
\examples{

run_cli(c("plot_alluvial", "--help"))

}
