% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python_env.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Setup Python environment for this package}
\usage{
setup_python_env(
  envname = "alluvialmatch_env",
  packages = c("numpy==1.23.5", "splitspy")
)
}
\description{
This function installs Miniconda if needed, creates a dedicated environment,
and installs the required Python packages.
}
