% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R
\name{data_color}
\alias{data_color}
\title{Color alluvia}
\usage{
data_color(
  data,
  cols,
  wt = NULL,
  method = "advanced",
  resolution = 1,
  verbose = FALSE,
  options = data_color_options()
)
}
\arguments{
\item{data}{A data frame, tibble, or CSV file path. Must be in one of two formats:
(1) wt == NULL: Each row represents an entity, each column represents a grouping, and each entry represents the membership of the entity in that row to the grouping in that column. Must contain at least two columns (two cols).
(2) wt != NULL: Each row represents a combination of groupings, each column from \code{cols} represents a grouping, and the column \code{wt} represents the number of entities in that combination of groupings. Must contain at least three columns (two \code{cols}, one \code{wt}).}

\item{cols}{Character vector. Vector of column names from \code{data} to be used in graphing (i.e., alluvial plotting).}

\item{wt}{Optional character. Column name from \code{data} that contains the weights of each combination of groupings if \code{data} is in format (2) (see above).}

\item{method}{Character. Matching colors methods. Choices are 'advanced' (default), 'none', 'left', 'right', or any value in \code{cols}.}

\item{resolution}{Numeric If \code{method == 'advanced'}, then choose resolution for the graph clustering algorithm.}

\item{verbose}{Logical. If TRUE, will display messages during the function.}

\item{options}{Additional arguments. See data_color_options}
}
\value{
A data frame where each row represents a combination of groupings, each column from \code{cols} represents a grouping, and the column \code{wt} ('value' if \code{wt} == NULL) represents the number of entities in that combination of groupings. There will be an additional column 'node_color'.
}
\description{
Colors a dataframe with the algorithm specified by \code{method}.
}
\examples{
# Example 1: data format 1
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
clus_df_gather <- data_preprocess(data, cols = c("method1", "method2"))

# Example 2: data format 2
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
clus_df_gather <- data |>
    dplyr::mutate_if(is.numeric, function(x) factor(x, levels = as.character(sort(unique(x))))) |>
    dplyr::group_by_all() |>
    dplyr::count(name = "value")
clus_df_gather <- data_preprocess(
    clus_df_gather,
    cols = c("method1", "method2"),
    wt = "value"
)

}
