% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R
\name{get_lode_clusters_options}
\alias{get_lode_clusters_options}
\title{Control Options for \code{get_lode_clusters()}}
\usage{
get_lode_clusters_options(
  method_advanced_option = c("leiden", "louvain"),
  cutoff = 0.5,
  preprocess_data = TRUE,
  print_params = FALSE
)
}
\arguments{
\item{method_advanced_option}{Character. When
\code{method == "advanced"}, selects the clustering algorithm.
Options: \code{"leiden"} (default) or \code{"louvain"}.}

\item{cutoff}{Numeric. If using a non-advanced algorithm, similarity score
threshold for deciding whether to reuse a color or assign a new one.}

\item{preprocess_data}{Logical. If TRUE (default), preprocess data with
\code{prep_for_lodes()}.}

\item{print_params}{Logical. If TRUE, prints parameters during execution.}
}
\value{
A named list of control parameters for use in \code{get_lode_clusters()}.
}
\description{
Creates a list of control parameters that modify the behavior of
\code{get_lode_clusters()}. These options allow fine-grained control over the coloring
algorithm without cluttering the main function interface.
}
\examples{

data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
opts <- get_lode_clusters_options(cutoff = 0.3, method_advanced_option = "louvain")
mapping <- get_lode_clusters(data = data, cols = c('method1', 'method2'), options = opts)

}
