% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{data_preprocess}
\alias{data_preprocess}
\title{1 line description}
\usage{
data_preprocess(df, graphing_columns = NULL, column_weights = NULL)
}
\arguments{
\item{df}{A data frame, tibble, or CSV file path. Must contain at least two columns, each representing a clustering/grouping of the same entities (rows).}
}
\value{
A data frame where each row provides the weight in the 'value' column that connects a given combination of values in graphing_columns.
}
\description{
Multiline
Description
}
\examples{
\dontrun{
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
clus_df_gather <- data_preprocess(df, graphing_columns = c('method1', 'method2'))
}

}
