% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{data_sort}
\alias{data_sort}
\title{Sorts a dataframe (e.g., the output of )}
\usage{
data_sort(
  df,
  graphing_columns = NULL,
  column_weights = NULL,
  sorting_algorithm = "neighbornet",
  optimize_column_order = TRUE,
  fixed_column = 1,
  random_initializations = 1,
  set_seed = 42,
  output_df_path = NULL
)
}
\arguments{
\item{df}{A data frame, tibble, or CSV file path. Must contain at least two columns, each representing a clustering/grouping of the same entities (rows).}
}
\value{
A \code{ggplot2} object representing the alluvial plot.
}
\description{
Multiline
Description
}
\examples{
\dontrun{
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
plot_alluvial(df)
}

}
