% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_calculation.R
\name{determine_crossing_edges}
\alias{determine_crossing_edges}
\title{Determine overlapping edges of graph run through data_sort}
\usage{
determine_crossing_edges(
  df,
  graphing_columns = NULL,
  column1 = NULL,
  column2 = NULL,
  column_weights = "value",
  minimum_edge_weight = 0,
  output_df_path = NULL,
  output_lode_df_path = NULL,
  return_weighted_layer_free_objective = FALSE
)
}
\arguments{
\item{df}{A data frame, tibble, or CSV file path. Must contain at least two columns, each representing a clustering/grouping of the same entities (rows).}

\item{column1}{Character. Name of the first column to plot. Optional if \code{df} has exactly two columns, or if \code{df} has exactly three columns including \code{column_weights}.}

\item{column2}{Character. Name of the second column to plot. Optional if \code{df} has exactly two columns, or if \code{df} has exactly three columns including \code{column_weights}}

\item{column_weights}{Optional numeric vector of weights (same length as number of rows in \code{df}) to weight each row differently when calculating flows.}

\item{minimum_edge_weight}{Optional positive integer that represents the minimum edge weight to count an edge in the calculation.}

\item{output_df_path}{Character. File path to save the dataframe (e.g., "df.csv"). If \code{NULL}, the dataframe is not saved.}

\item{return_weighted_layer_free_objective}{Bool Whether to return a list of overlapping edges (FALSE) or the sum of products of overlapping edges (TRUE)}
}
\value{
A data frame.
}
\description{
Determine overlapping edges of graph run through data_sort
}
\examples{
\dontrun{
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
df <- data_sort(df)
result <- determine_crossing_edges(df, column1 = "col1_int", column2 = "col2_int")
}

}
