% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{determine_weighted_layer_free_objective}
\alias{determine_weighted_layer_free_objective}
\title{Determine sum of products of overlapping edges}
\usage{
determine_weighted_layer_free_objective(
  crossing_edges,
  minimum_edge_weight = 0
)
}
\arguments{
\item{crossing_edges}{A CSV path or list as outputted from determine_crossing_edges.}

\item{minimum_edge_weight}{Optional positive integer that represents the minimum edge weight to count an edge in the calculation.}
}
\value{
A data frame.
}
\description{
Determine sum of products of overlapping edges
}
\examples{
\dontrun{
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
greedy_wolf(df)
crossing_edges <- determine_crossing_edges(df, column1 = "method1", column2 = "method2")
objective <- determine_weighted_layer_free_objective(crossing_edges)
}

}
