% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{greedy_wolf}
\alias{greedy_wolf}
\title{Perform a greedy algorithm heuristic for the Weighted One Layer Free problem}
\usage{
greedy_wolf(
  df,
  column1 = NULL,
  column2 = NULL,
  column_weights = NULL,
  sorting_algorithm = "greedy_WBLF",
  fixed_column = 1,
  random_initializations = 1,
  set_seed = 42,
  output_df_path = NULL
)
}
\arguments{
\item{df}{A data frame, tibble, or CSV file path. Must contain at least two columns, each representing a clustering/grouping of the same entities (rows).}

\item{column1}{Character. Name of the first column to plot. Optional if \code{df} has exactly two columns, or if \code{df} has exactly three columns including \code{column_weights}.}

\item{column2}{Character. Name of the second column to plot. Optional if \code{df} has exactly two columns, or if \code{df} has exactly three columns including \code{column_weights}}

\item{column_weights}{Optional numeric vector of weights (same length as number of rows in \code{df}) to weight each row differently when calculating flows.}

\item{fixed_column}{Optional Character or Integer. Name of the column to fix, if desiring a one-layer free algorithm. If NULL, then implement both layers free. If 1, then will fix column1; if 2, then will fix column2}

\item{random_initializations}{Optional Integer. Number of random initializations of the WLF heuristic to perform.}

\item{output_df_path}{Character. File path to save the dataframe (e.g., "df.csv"). If \code{NULL}, the dataframe is not saved.}
}
\value{
A data frame.
}
\description{
Perform a greedy algorithm heuristic for the Weighted One Layer Free problem
}
\examples{
\dontrun{
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
greedy_wolf(df)
}

}
