% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R
\name{lode_cluster_pal}
\alias{lode_cluster_pal}
\title{Make stratum color list}
\usage{
lode_cluster_pal(data, cols, mapping, color_palette = NULL)
}
\arguments{
\item{data}{A data frame with columns corresponding to axes, with factors indicating sorting. (eg run through sort_to_uncross)}

\item{cols}{Character vector. Vector of column names from \code{data} to be used in graphing (i.e., alluvial plotting).}

\item{mapping}{List. Output from get_lode_clusters.}

\item{color_palette}{Optional named list or vector mapping values in the graphing columns to colors. Overrides default palette.}
}
\value{
A vector of colors.
}
\description{
Convert color mapping made by get_lode_clusters into list for scale_fill_manual
}
\examples{
# Example 1
data <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
cols = c("method1", "method2")
clus_df_gather <- sort_to_uncross(data, cols = cols, method = "tsp")
color_mapping <- get_lode_clusters(data = clus_df_gather, cols = cols)
color_list <- lode_cluster_pal(data = clus_df_gather, cols = cols, mapping = color_mapping)

}
