% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{plot_alluvial}
\alias{plot_alluvial}
\title{Generate an Alluvial Plot with Minimal Cluster Cross-over}
\usage{
plot_alluvial(
  df,
  column1 = NULL,
  column2 = NULL,
  fixed_column = 1,
  sorting_algorithm = "greedy_WBLF",
  random_initializations = 1,
  color_list = NULL,
  color_boxes = TRUE,
  color_bands = FALSE,
  color_band_list = NULL,
  color_band_column = NULL,
  color_band_boundary = FALSE,
  match_colors = TRUE,
  alluvial_alpha = 0.5,
  include_labels_in_boxes = TRUE,
  include_axis_titles = TRUE,
  include_group_sizes = TRUE,
  column_weights = NULL,
  output_plot_path = NULL,
  output_df_path = NULL,
  return_greedy_wolf = FALSE,
  set_seed = 42
)
}
\arguments{
\item{df}{A data frame, tibble, or CSV file path. Must contain at least two columns, each representing a clustering/grouping of the same entities (rows).}

\item{column1}{Character. Name of the first column to plot. Optional if \code{df} has exactly two columns, or if \code{df} has exactly three columns including \code{column_weights}.}

\item{column2}{Character. Name of the second column to plot. Optional if \code{df} has exactly two columns, or if \code{df} has exactly three columns including \code{column_weights}.}

\item{fixed_column}{Character. Name of the column to fix, if desiring a one-layer free algorithm. If NULL, then implement both layers free.}

\item{sorting_algorithm}{Character. Must be greedy_WBLF, greedy_WOLF, or None.}

\item{random_initializations}{Optional Integer. Number of random initializations of the WLF heuristic to perform.}

\item{color_list}{Optional named list or vector of colors to override default group colors.}

\item{color_boxes}{Logical. Whether to color the rectangular strata boxes representing groups.}

\item{color_bands}{Logical. Whether to color the alluvial bands connecting the groups.}

\item{color_band_list}{Optional named list or vector of colors to override default band colors.}

\item{color_band_column}{Optional Character. Which column to use for coloring bands.}

\item{color_band_boundary}{Logical. Whether or not to color boundaries between bands}

\item{match_colors}{Logical. If \code{TRUE}, assigns consistent colors between column1 and column2 where matched.}

\item{alluvial_alpha}{Numeric between 0 and 1. Transparency level for the alluvial bands.}

\item{include_labels_in_boxes}{Logical. Whether to include text labels inside the rectangular group boxes.}

\item{include_axis_titles}{Logical. Whether to display axis titles for column1 and column2.}

\item{include_group_sizes}{Logical. If \code{TRUE}, includes group sizes in the labels (e.g., "Group A (42)").}

\item{column_weights}{Optional numeric vector of weights (same length as number of rows in \code{df}) to weight each row differently when calculating flows.}

\item{output_plot_path}{Character. File path to save the plot (e.g., "plot.png"). If \code{NULL}, the plot is not saved.}

\item{output_df_path}{Character. File path to save the dataframe (e.g., "df.csv"). If \code{NULL}, the dataframe is not saved.}

\item{set_seed}{Optional Integer. Seed for random initializations of the WLF heuristic to perform.}
}
\value{
A \code{ggplot2} object representing the alluvial plot.
}
\description{
Creates a two-axis alluvial plot to visualize the relationship between two categorical groupings (e.g., cluster assignments from different methods),
minimizing crossover and aligning clusters based on agreement.
}
\examples{
\dontrun{
df <- data.frame(method1 = sample(1:3, 100, TRUE), method2 = sample(1:3, 100, TRUE))
plot_alluvial(df)
}

}
