% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python_env.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Setup Python environment for this package}
\usage{
setup_python_env(
  envname = "wompwomp_env",
  packages = c(numpy = "numpy==1.23.5", splitspy = "splitspy", pandas = "pandas", scipy =
    "scipy", leidenalg = "leidenalg", igraph = "python-igraph"),
  use_conda = TRUE,
  yes = FALSE
)
}
\arguments{
\item{envname}{Conda environment name}

\item{packages}{Packages to install}

\item{use_conda}{Use conda vs. virtualenv for python environment}
}
\value{
Invisibly returns TRUE if the environment is set up successfully.
}
\description{
This function installs Miniconda if needed, creates a dedicated environment, and installs the required Python packages.
}
\examples{
\dontrun{
setup_python_env()
}

}
