% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_alluvial.R
\name{sort_to_uncross_options}
\alias{sort_to_uncross_options}
\title{Control Options for \code{sort_to_uncross()}}
\usage{
sort_to_uncross_options(
  optimize_column_order_per_cycle = FALSE,
  matrix_initialization_value = 1e+06,
  same_side_matrix_initialization_value = 1e+06,
  matrix_initialization_value_column_order = 1e+06,
  weight_scalar_column_order = 1,
  column_metric = c("edge_crossing", "ari"),
  weighted_metric = TRUE,
  cycle_start_positions = NULL,
  random_initializations = 1,
  preprocess_data = TRUE,
  default_sorting = c("alphabetical", "reverse_alphabetical", "increasing", "decreasing",
    "random", "fixed"),
  print_params = FALSE,
  do_compute_alluvial_statistics = FALSE
)
}
\arguments{
\item{optimize_column_order_per_cycle}{Logical. If TRUE, will optimize the order of \code{cols} to minimize edge overlap upon each cycle. If FALSE, will optimize the order of \code{cols} to minimize edge overlap on the beginning cycle only. Only applies when \code{method \%in\% c('neighbornet', 'tsp')} and \code{length(cols) > 2}.}

\item{matrix_initialization_value}{Positive integer. Initialized value in distance matrix for nodes in different layers without a shared edge/path. Only applies when \code{method \%in\% c('neighbornet', 'tsp')}.}

\item{same_side_matrix_initialization_value}{Positive integer. Initialized value in distance matrix for nodes in the same layer. Only applies when \code{method \%in\% c('neighbornet', 'tsp')}.}

\item{matrix_initialization_value_column_order}{Positive integer. Initialized value in distance matrix for optimizing column order. Only applies when \code{column_method != 'none'}.}

\item{weight_scalar_column_order}{Positive integer. Scalar with which to loss function after taking their log1p in the distance matrix for optimizing column order. Only applies when \code{column_method != 'none'}.}

\item{column_metric}{Character. Metric to use for determining column order. Options are "edge_crossing" (default) or "ARI". Only applies when \code{column_method != 'none'}.}

\item{weighted_metric}{Logical. Determines if the objective is total number of edge crossings (weighted_metric=FALSE) or sum of product of overlapping edge weights (weighted_metric=TRUE).}

\item{cycle_start_positions}{Set. Cycle start positions to consider. Anything outside this set will be skipped. Only applies when \code{method \%in\% c('neighbornet', 'tsp')}.}

\item{random_initializations}{Integer. Number of random initializations for the positions of each grouping in \code{cols}. Only applies when \code{method == 'greedy_wolf' or method == 'greedy_wblf'}.}

\item{preprocess_data}{Logical. If TRUE, will preprocess the data with the \code{\link[=prep_for_lodes]{prep_for_lodes()}} function.}

\item{default_sorting}{Character. Default column sorting in \code{\link[=prep_for_lodes]{prep_for_lodes()}} if integer columns do not exist. Options are 'alphabetical' (default), 'reverse_alphabetical', 'increasing', 'decreasing', 'random'.}

\item{print_params}{Logical. If TRUE, will print function params.}

\item{do_compute_alluvial_statistics}{Internal flag; not recommended to modify.}
}
\value{
A named list of control parameters, to be passed into \code{\link[=sort_to_uncross]{sort_to_uncross()}}
via the \code{options} argument.
}
\description{
Creates a list of control parameters that modify the behavior of
\code{\link[=sort_to_uncross]{sort_to_uncross()}}. These options allow tuning algorithmic behavior without
cluttering the main function arguments.
}
\examples{
data <- data.frame(
  method1 = LETTERS[sample(1:3, 100, TRUE)],
  method2 = LETTERS[27 - sample(1:3, 100, TRUE)]
)
opts <- sort_to_uncross_options(
  default_sorting = "reverse_alphabetical",
  matrix_initialization_value = 100
)
sort_to_uncross(data = data, cols = c('method1', 'method2'), options = opts)

}
