% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R, R/objective_calculation.R,
%   R/sort_alluvial.R, R/utils.R, R/wompwomp-package.R
\docType{package}
\name{wompwomp-imports}
\alias{wompwomp-package}
\alias{wompwomp-imports}
\alias{wompwomp}
\title{wompwomp: Cluster-matching alluvial plots}
\description{
Sort k-partite graphs with node order, layer order, and node grouping optimized with a heuristic to (nearly) minimize edge crossings. Useful for improving visualizations with alluvial plots by "untangling" the graphs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachterlab/wompwomp}
  \item Report bugs at \url{https://github.com/pachterlab/wompwomp/issues}
}

}
\author{
\strong{Maintainer}: Joseph Rich \email{jmrich@caltech.edu} (\href{https://orcid.org/0000-0003-1400-8479}{ORCID})

Authors:
\itemize{
  \item Conrad Oakes \email{coakes@caltech.edu}
  \item Lior Pachter (\href{https://orcid.org/0000-0002-9164-6231}{ORCID})
}

}
\keyword{internal}
