% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R, R/objective_calculation.R,
%   R/sort_alluvial.R, R/utils.R
\docType{package}
\name{wompwomp}
\alias{wompwomp}
\title{wompwomp: Coloring alluvia}
\description{
Coloring alluvia

Objective calculation

Sorting alluvia

Utils
}
