% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_alluvial.R, R/objective_calculation.R,
%   R/sort_alluvial.R, R/utils.R, R/wompwomp-package.R
\docType{package}
\name{wompwomp-imports}
\alias{wompwomp-package}
\alias{wompwomp-imports}
\alias{wompwomp}
\title{wompwomp: Cluster-matching alluvial plots}
\description{
Alluvial plots are great for visualizing the concordance between two different clustering methods applied to a dataset. However, these plots only allow for cluster agreement when each side of the bipartite graph is lined up in such a way that there is minimal cross-over between clusters, thus allowing one to visualize something analogous to an Adjusted Rand Index. wompwomp utilizes ggalluvial and igraph::max_bipartite_match in order to achieve this goal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachterlab/wompwomp}
  \item Report bugs at \url{https://github.com/pachterlab/wompwomp/issues}
}

}
\author{
\strong{Maintainer}: Joseph Rich \email{jmrich@caltech.edu} (\href{https://orcid.org/0000-0003-1400-8479}{ORCID})

Authors:
\itemize{
  \item Conrad Oakes \email{coakes@caltech.edu}
  \item Lior Pachter (\href{https://orcid.org/0000-0002-9164-6231}{ORCID})
}

}
\keyword{internal}
