% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_calculation.R, R/plot_alluvial.R
\docType{package}
\name{wompwomp}
\alias{wompwomp}
\title{wompwomp: Cluster-matching alluvial plots}
\description{
Main plotting function and helpers for bipartite-matching-based alluvial diagrams

Main plotting function and helpers for bipartite-matching-based alluvial diagrams
}
